+++ src/bun.js/bindings/webcore/AbortSignal.h
@@ -63,7 +63,7 @@ class AbortSignal final : public RefCounted<AbortSignal>, public EventTargetWith
     WTF_MAKE_TZONE_ALLOCATED(AbortSignal);
 
 private:
-    enum class AbortSignalFlags : uint32_t {
+    enum class AbortSignalFlags : uint8_t {
         Dependent = 1,
         Aborted = 2,
         HasAbortEventListener = 4,
@@ -86,7 +86,7 @@ public:
     void signalAbort(JSC::JSValue reason);
     void signalFollow(AbortSignal&);
 
-    bool aborted() const { return m_flags & static_cast<uint32_t>(AbortSignalFlags::Aborted); }
+    bool aborted() const { return m_flags & static_cast<uint8_t>(AbortSignalFlags::Aborted); }
     void markAborted(JSC::JSValue reason);
     void runAbortSteps();
 
@@ -98,8 +98,8 @@ public:
     void addNativeCallback(NativeCallbackTuple callback) { m_native_callbacks.append(callback); }
 
     bool hasActiveTimeoutTimer() const { return m_timeout != nullptr; }
-    bool hasAbortEventListener() const { return m_flags & static_cast<uint32_t>(AbortSignalFlags::HasAbortEventListener); }
-    bool isFiringEventListeners() const { return m_flags & static_cast<uint32_t>(AbortSignalFlags::IsFiringEventListeners); }
+    bool hasAbortEventListener() const { return m_flags & static_cast<uint8_t>(AbortSignalFlags::HasAbortEventListener); }
+    bool isFiringEventListeners() const { return m_flags & static_cast<uint8_t>(AbortSignalFlags::IsFiringEventListeners); }
 
     using RefCounted::deref;
     using RefCounted::ref;
@@ -120,7 +120,7 @@ public:
     void incrementPendingActivityCount() { ++pendingActivityCount; }
     void decrementPendingActivityCount() { --pendingActivityCount; }
     bool hasPendingActivity() const { return pendingActivityCount > 0; }
-    bool isDependent() const { return m_flags & static_cast<uint32_t>(AbortSignalFlags::Dependent); }
+    bool isDependent() const { return m_flags & static_cast<uint8_t>(AbortSignalFlags::Dependent); }
 
     size_t memoryCost() const;
 
@@ -136,37 +136,37 @@ private:
     void addDependentSignal(AbortSignal&);
     void cancelTimer();
 
-    void applyFlags(uint32_t flags) { m_flags |= flags; }
+    void applyFlags(uint8_t flags) { m_flags |= flags; }
     void setIsDependent(bool isDependent)
     {
         if (isDependent) {
-            m_flags |= static_cast<uint32_t>(AbortSignalFlags::Dependent);
+            m_flags |= static_cast<uint8_t>(AbortSignalFlags::Dependent);
         } else {
-            m_flags &= ~static_cast<uint32_t>(AbortSignalFlags::Dependent);
+            m_flags &= ~static_cast<uint8_t>(AbortSignalFlags::Dependent);
         }
     }
     void setAborted(bool aborted)
     {
         if (aborted) {
-            m_flags |= static_cast<uint32_t>(AbortSignalFlags::Aborted);
+            m_flags |= static_cast<uint8_t>(AbortSignalFlags::Aborted);
         } else {
-            m_flags &= ~static_cast<uint32_t>(AbortSignalFlags::Aborted);
+            m_flags &= ~static_cast<uint8_t>(AbortSignalFlags::Aborted);
         }
     }
     void setHasAbortEventListener(bool hasAbortEventListener)
     {
         if (hasAbortEventListener) {
-            m_flags |= static_cast<uint32_t>(AbortSignalFlags::HasAbortEventListener);
+            m_flags |= static_cast<uint8_t>(AbortSignalFlags::HasAbortEventListener);
         } else {
-            m_flags &= ~static_cast<uint32_t>(AbortSignalFlags::HasAbortEventListener);
+            m_flags &= ~static_cast<uint8_t>(AbortSignalFlags::HasAbortEventListener);
         }
     }
     void setIsFiringEventListeners(bool isFiringEventListeners)
     {
         if (isFiringEventListeners) {
-            m_flags |= static_cast<uint32_t>(AbortSignalFlags::IsFiringEventListeners);
+            m_flags |= static_cast<uint8_t>(AbortSignalFlags::IsFiringEventListeners);
         } else {
-            m_flags &= ~static_cast<uint32_t>(AbortSignalFlags::IsFiringEventListeners);
+            m_flags &= ~static_cast<uint8_t>(AbortSignalFlags::IsFiringEventListeners);
         }
     }
 
@@ -187,7 +187,7 @@ private:
     std::atomic<uint32_t> pendingActivityCount { 0 };
     uint32_t m_algorithmIdentifier { 0 };
     AbortSignalTimeout m_timeout { nullptr };
-    uint32_t m_flags { 0 };
+    uint8_t m_flags { 0 };
 };
 
 WebCoreOpaqueRoot root(AbortSignal*);
